Fix assuming master rank is 0 for broadcast root
In the constructor the master process was responsible for initialising
the pos, vel, mass fields. In the broadcast, the root rank=0 is used,
while the master process world rank is 0, it is just safer to choose
the process with rank 0 to initialise the fields, instead of relying on
isMaster. The member variable 'isMaster' will likely be removed later.